+++ docs/source/conf.py
@@ -26,6 +26,7 @@ project = "sam"
 copyright = "2018-2022, Royal HaskoningDHV"
 author = "Royal HaskoningDHV"
 
+
 config = toml.load("../../pyproject.toml")
 
 # The short X.Y version